+++ FlowCrypt/Functionality/Api/Remote Pub Key Apis/AttesterApi.swift
@@ -34,7 +34,7 @@ final class AttesterApi: AttesterApiType {
 
     private func constructUrlBase() -> String {
         guard !Bundle.isDebugBundleWithArgument("--mock-attester-api") else {
-            return "https://127.0.0.1:8001/attester" // mock
+            return "\(GeneralConstants.Mock.backendUrl)attester" // mock
         }
         return "https://flowcrypt.com/attester" // live
     }
