+++ FlowCrypt/Functionality/Mail Provider/Gmail/GmailService.swift
@@ -20,8 +20,8 @@ class GmailService: MailServiceProvider {
     var gmailService: GTLRService {
         let service = GTLRGmailService()
 
-        if Bundle.isDebugBundleWithArgument("--mock-gmail-api") {
-            service.rootURLString = "https://127.0.0.1:8001/"
+        if Bundle.isMockGmailApi {
+            service.rootURLString = GeneralConstants.Mock.backendUrl
         }
 
         if gmailUserService.authorization == nil {
