--- appium/api-mocks/mock-ssl-cert/cert.pem
@@ -1,21 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDhjCCAm6gAwIBAgIJAM8czFmkbKHvMA0GCSqGSIb3DQEBCwUAMHgxCzAJBgNV
-BAYTAlhYMQwwCgYDVQQIDANOL0ExDDAKBgNVBAcMA04vQTEgMB4GA1UECgwXU2Vs
-Zi1zaWduZWQgY2VydGlmaWNhdGUxKzApBgNVBAMMIjEyNy4wLjAuMTogU2VsZi1z
-aWduZWQgY2VydGlmaWNhdGUwHhcNMjIwNjA5MDY0MjQ2WhcNMjQwNjA4MDY0MjQ2
-WjB4MQswCQYDVQQGEwJYWDEMMAoGA1UECAwDTi9BMQwwCgYDVQQHDANOL0ExIDAe
-BgNVBAoMF1NlbGYtc2lnbmVkIGNlcnRpZmljYXRlMSswKQYDVQQDDCIxMjcuMC4w
-LjE6IFNlbGYtc2lnbmVkIGNlcnRpZmljYXRlMIIBIjANBgkqhkiG9w0BAQEFAAOC
-AQ8AMIIBCgKCAQEAvgozjkzUVQ26zOT8BBXE+VMXJdiZowygNnGH3HxN0J9jU343
-fH0R1fm6HA98AA7NXJnb+RIjaAk/vub0eurn3HiaCFlMMSa+C/jdZg/4vQXLao6H
-cfDNKR5hz4VvGVFQ2kWr4GYZKm6Iyqq4aYmdjAx7zbvpocWyHWBSu75Y8mmk5YJM
-jr+47e//uGTDdVgcrx/KLGtTi44wV8ZkC5mr58TaqHI+y4aTvPrCF05+XMx8O8jf
-MbgOkagi3TVAlXlrqHh1PZbaRjIb7qNJ22eOJ4qchQ7GFyqGcmuKgVcv1xS0zrEP
-w9E5G4bFogjFdbm5tHkzRifb0fsh0XDlClSHeQIDAQABoxMwETAPBgNVHREECDAG
-hwR/AAABMA0GCSqGSIb3DQEBCwUAA4IBAQBjFEwVXWHpJordck5ritnxNAEvGGN0
-2+7idBKEQCuTljjEr0QuU3zcQmjhO0BLwYMJUHWXqvCXuCSvc15EiEQEDrzd2sWd
-UtZEozy0aV4nRoCKUqDy2pcGsmXiskAs4J7vYyB97a0DNA/99zhV99UyF49tNSy4
-nYmIb6CGOY8LyxqNdNiul4b9d0p9LVJtICNHF4AQ3ovalECL7Inrat4p03JNruNp
-awhZb/Ng8hIidke+oZ0se3kekzA7bMMcxkYGwI38QVwKchRE97G0Fmcgpbm+UQzs
-MvlNXquf2JoVtgkYoXcrxC/fKGXMKdXSGfMzrMb8EBnxkiuXr/pqNNAb
------END CERTIFICATE-----
