--- appium/api-mocks/mock-ssl-cert/cert.pem.mock
@@ -1,19 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDAjCCAeqgAwIBAgIJALr1VAaYpiR2MA0GCSqGSIb3DQEBBQUAMCkxJzAlBgNV
-BAMTHmdvb2dsZS5tb2NrLmZsb3djcnlwdGxvY2FsLmNvbTAeFw0yMDA1MjUxODMw
-MjZaFw0zMDA1MjMxODMwMjZaMCkxJzAlBgNVBAMTHmdvb2dsZS5tb2NrLmZsb3dj
-cnlwdGxvY2FsLmNvbTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAK8r
-E2xAE65Wcu1hHM3eNnyqZ+6i4R/V3KpRw+QOPgyB6uSIYiE0uaYQlx1Zcu5hhOX0
-XmjJuM7oq1fy3nKktrkg0wCyZKg04S+MyNclFzBP4RSbDb8e0vbCnLwYl4hg/247
-Jz/NSqKekXY9fWjX3X7NcYYKrIrC+vNBmfNnZ2KsGkITe2gkXxuoa7C3CP2LrIhH
-4FQl38miLJYmGxPxARfIuXsB0PY7I4FM1GaZ131rCLBhkEC+mjsou0kHmKpLc/sh
-7QIhQG38QyUUHnL/RLjGW1smz77vxLYC1GVCS1IBQcce3OBH9WJWHV+ewOXwTaVp
-5C0SB3vb7rXsLaCDuMMCAwEAAaMtMCswKQYDVR0RBCIwIIIeZ29vZ2xlLm1vY2su
-Zmxvd2NyeXB0bG9jYWwuY29tMA0GCSqGSIb3DQEBBQUAA4IBAQBk5AGIu26GwTdt
-JryCjrMTKfO3eLZKfwd9TrZujjVC/1G2auF+3RHD9QFGzwrLhFrAQSjwNrYdm2KA
-k1qfKRAjf8IE1KKj7sqJfNUeVObk75gHBByObcnyrhZl58j6bYvgrrtIGZZEado0
-BIlhhYLLMDpGGTpzpVZ7Zw2huTVgP4QGKwycumHTSFYCqZ9k36CvmDZCBXxH3Gyb
-m18aiZEoXsorzzmTCLqpHOVGnSfmeNTLEE2PxlAL0qv212XpJVw/dNQtVzDdj9nr
-zRf6xdAIGUYhZfBkPKilale8XG88NA9efRxFdA+B5D+/ecX94Bv3hL1xHP229Fd+
-3HtRLE/r
------END CERTIFICATE-----
