--- appium/api-mocks/mock-ssl-cert/key.pem
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQC+CjOOTNRVDbrM
-5PwEFcT5Uxcl2JmjDKA2cYfcfE3Qn2NTfjd8fRHV+bocD3wADs1cmdv5EiNoCT++
-5vR66ufceJoIWUwxJr4L+N1mD/i9Bctqjodx8M0pHmHPhW8ZUVDaRavgZhkqbojK
-qrhpiZ2MDHvNu+mhxbIdYFK7vljyaaTlgkyOv7jt7/+4ZMN1WByvH8osa1OLjjBX
-xmQLmavnxNqocj7LhpO8+sIXTn5czHw7yN8xuA6RqCLdNUCVeWuoeHU9ltpGMhvu
-o0nbZ44nipyFDsYXKoZya4qBVy/XFLTOsQ/D0TkbhsWiCMV1ubm0eTNGJ9vR+yHR
-cOUKVId5AgMBAAECggEADX1T44n4mID8/YZTjUauULRhNEKHxVckGDk+15vLOj/W
-IRsYfTGzjmjHmS5QMwDZ+vTA9JEhCkV9zOM//2mlQ7Ub2fPcwLBrVRY5QQFpOqGq
-qsmrO59IMN8DOmPSWw22m4fOh4M3zhdChoBPzs4B7OsF7rikAOOznPo1ZTZBwhM2
-WgZeTMDx7hzZavI8JoU8YsroTyIZfB28QemzbwwE4oU7NJx78M9S4wRLO76GP0N3
-tCAC8Q6esDDWMUnDsfAehhecRKNNuEGyVYdZK7FwGda3VMyZWw0qnlD8Xj1jyM6F
-P278WVV2r/4Dmrvrw2LCNVl7P3Bi47IpETZgi5GAMQKBgQDiqVXQE5zkPDhTGmKB
-WviWuqweB0FhJidLgUc4/+MfK6h2/7MKN4/9weP0AzMDbF84fQG9k93XDXIEQYK1
-AKsU8jYeh+YLV7hCrhyfNq7uWkyQyyAad+WgM5HHnZUS6pXjaU9FdpRZ1CMm/BhZ
-NeUyLWGNF5UnE93GhyJgy/vj4wKBgQDWo17gPmykwoOIFqdl8OoKBopk7xoQbj/p
-etGjoWzQLAkk0s6VXBfXOk8IMC+fKR+edQ2fBsOO+D2xT8Ieo1gb0W/GE7SkCEMd
-wIsj+4Zyt4P6zb851tIf8gzmz9nGS65AOqpxIqehH/LwtkER0Wk+OL8gvQLW21U5
-pXeVFfOd8wKBgFxkHgCVHgzCp0ropGKKDG6UZPhQiMvq23WhmsCXPaNFDoSDMh3m
-DCGp2Rtr+m1YzW+fNQhrr6+rYWBuu7LANPxyxrCnQJQ7C3bStgQILEIoOLGEuu9P
-FCC7JJ8zksKFMNfj41YzQcFkAXgwplQm7egIK41xoXCuXmZASzkdrO11AoGBAM4C
-I+/rOm+w2nnl9Gmdn6Rbvr092LBogSNhZJE/QKdVfLlBNFKa/67uIijjxvFmkRiP
-lLFL1qsEjIIA4XpyoQCiNjvRchH2dCS8R1AT/GlixmQU0jN7glZzBfsLCCSRQwAW
-zryj3GyDs7Nh2nZ1niPAZ4HOqEj8X4Ji1LHdTeQNAoGBALa26WCEUKFWflw4x6nS
-6w+0mqbx1bu7kv7YbqOHRvVa80Kgu1zoXWKL+wmInU1KqOWnpTd3G9wACM69OL3j
-g3tZSszgepDlV5S/ZeHfGA2ZF3Jq71EPRneQMWn6vFJmJoKEE2Dijpm97lkanT5z
-iGACPFYieSRGc6QUHPyasgiw
------END PRIVATE KEY-----
