--- appium/api-mocks/mock-ssl-cert/key.pem.mock
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEogIBAAKCAQEArysTbEATrlZy7WEczd42fKpn7qLhH9XcqlHD5A4+DIHq5Ihi
-ITS5phCXHVly7mGE5fReaMm4zuirV/LecqS2uSDTALJkqDThL4zI1yUXME/hFJsN
-vx7S9sKcvBiXiGD/bjsnP81Kop6Rdj19aNfdfs1xhgqsisL680GZ82dnYqwaQhN7
-aCRfG6hrsLcI/YusiEfgVCXfyaIsliYbE/EBF8i5ewHQ9jsjgUzUZpnXfWsIsGGQ
-QL6aOyi7SQeYqktz+yHtAiFAbfxDJRQecv9EuMZbWybPvu/EtgLUZUJLUgFBxx7c
-4Ef1YlYdX57A5fBNpWnkLRIHe9vutewtoIO4wwIDAQABAoIBAGglb03cJxdvVLNo
-Gs0XwkOKuXJ7qufXqnQlWxeblC6xi6vGnrlpdLuUf5Q2m4u5tV9XVE5vKQ7z7PPE
-dZep8O3O40U1y3zqu7wes0afSdMymvs6WQj51apa+P8oi+fM6/FpACQB9c8MAiZP
-1uFf/tQmZ8mvTPjPho7cwSCjqSdTeb7jubCEmNdhQcYqTY8la1LQ1fxi6wpHbpnJ
-iw+HyfApQcLlf4mVP04P+1SVsuliJi5VAqcpn/OPFVX4KMATkSNM0a6doNBVonAU
-MVu6Wor9r85z1R4gzaQms3Ggs6onwr+IhZfDuZ6fO4pKZar3rycWXpnwKvT6AuJe
-a5qYwjkCgYEA3w4BIhkaqAYOnJZYEWYkyPTuBaoT8Zqg0FeGahY/1+36nTSKfAPr
-Zs1DZbKJiFZiT7bOIpwpIqdT+R6Qviw88gss4xIL7zUT4MXBcYtXD0GrCass6xpO
-cxaUg81ioI+r7wxCvDvOlwKFBvfRRCZZ9T88nXQ523vB+HaSF5ef8QcCgYEAyQpr
-3iwXGjt9RkRR5QjnxQD7x5cVWBB9uy+tFP+/2tRcaskl9aQ/h5iPKYR/KVkhQf7i
-GPlk5qKxM8stI9M/bSaWRfVevaJTtphVUE0h8UycRr9YgrD0u6mw5dksBYmoXltc
-+I6ZuS66CEbLoNXCO2udBwRvC8e5dnIC2zuuF2UCgYAKsno9b2TG4Bt9SPyKQRLM
-jGqJ3QE92gOSWlA48D45ldEqDuYBUpLMiFWOIATWhAYM+qBiKokSf2HYla2tYSO/
-G+gYL7pgy3PbmMBROLSBJ84TFaF+2PEgNdoDhCsRtjIxsa3VpyyeM2G71tCzav9v
-covupS9NJl9EEjOpa/0gqwKBgAVmIj91L6mkr1iTvR9Jfe/sdtaFWC1lQ/si7KMD
-u0NFMLYtcLeIEayyjCWxSTjzC9GZ5riE+hC3jcq2SiAMLEfiIWb/j+vH0De3H5No
-fuDcDdZYjSf1WJ0xkW6axnPfnoXqLjT+n2I5vwU54X4QP4cY/L6Tf+fxFSn5Jdp9
-X8+BAoGAaYF/4tkfeiVzaVeinKs4MV22SJnh1Tjm9cNjOdX4++/b0IWwus42zh5g
-6OvYWB3OwiJiQny0pJUQlPMJjHxQ0FWcEriGqZkUqm2Hfla5whJAt6lt4AZuwte+
-aqn6RdxEZRP2sQFY3igZ8+TsxryV5wRIWbv9pVWfn0t8/WlXREI=
------END RSA PRIVATE KEY-----
